+++ package.json
@@ -25,10 +25,10 @@
     "build:ts": "yarn build:ts:cjs && yarn build:ts:esm",
     "build:ts:cjs": "tsc --project tsconfig.cjs.json",
     "build:ts:esm": "tsc --project tsconfig.esm.json",
-    "build:gen": "yarn -s ts-node generator/cli --download-templates-repo --generate-type-script --generate-migration-sql",
+    "build:gen": "yarn -s ts-node generator/cli --download-templates-repo --generate-type-script --generate-migration-scripts",
     "build:gen:download-templates-repo": "yarn -s ts-node generator/cli --download-templates-repo",
     "build:gen:generate-type-script": "yarn -s ts-node generator/cli --generate-type-script",
-    "build:gen:generate-migration-sql": "yarn -s ts-node generator/cli --generate-migration-sql",
+    "build:gen:generate-migration-scripts": "yarn -s ts-node generator/cli --generate-migration-scripts",
     "test": "jest",
     "test:unit": "yarn -s test --testPathIgnorePatterns e2e templates-repo",
     "test:e2e": "yarn -s test e2e",
